requirements.txt
streamlit
langchain
langchain-core
langgraph
langchain-google-genai
ipython
python-dotenv
langchain-community